repository factory class to get domain specific cake class
