+++ app/server/appsmith-server/src/main/java/com/appsmith/server/helpers/RepositoryFactory.java
@@ -1,0 +1,5 @@
+package com.appsmith.server.helpers;
+
+import com.appsmith.server.helpers.ce.RepositoryFactoryCE;
+
+public interface RepositoryFactory extends RepositoryFactoryCE {}
