+++ app/server/appsmith-server/src/main/java/com/appsmith/server/helpers/ce/RepositoryFactoryCE.java
@@ -1,0 +1,7 @@
+package com.appsmith.server.helpers.ce;
+
+import com.appsmith.server.repositories.AppsmithRepository;
+
+public interface RepositoryFactoryCE {
+    AppsmithRepository<?> getRepositoryObjectFromEntityType(Object object);
+}
